refactor: Move operations to `ops`
Move all operation mixins to the `ops` subpackage, mirroring the vaccel
C code structure

PR: https://github.com/nubificus/vaccel-python/pull/16
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
