+++ vaccel/ops/image.py
@@ -1,8 +1,8 @@
 """Image-related operations."""
 
-from ._c_types import CBytes
-from ._libvaccel import lib
-from .error import FFIError
+from vaccel._c_types import CBytes
+from vaccel._libvaccel import lib
+from vaccel.error import FFIError
 
 
 class ImageMixin:
