+++ vaccel/ops/minmax.py
@@ -1,8 +1,8 @@
 """Minmax operation."""
 
-from ._c_types import CBytes, CFloat
-from ._libvaccel import ffi, lib
-from .error import FFIError
+from vaccel._c_types import CBytes, CFloat
+from vaccel._libvaccel import ffi, lib
+from vaccel.error import FFIError
 
 
 class MinmaxMixin:
